SceneInspector : Add aggregate attribute inspectors
This allows us to show diff background colours at each parent level in the hierarchy. If this seems like a good direction to go in, then we should add aggregate inspectors for everything else, so you get diff breadcrumbs all the way down the hierarchy.

The biggest question if extending this to other sections is what value to use for Transform and Bound sections. We want to use the full transform/bound to show diffs appropriately, but I don't think we want any visible value to show up in the value column.

Another thing that bothers me a little bit is putting the number of attributes in a column labelled `Value`. It's not the value, it's just a summary. Is there some better way of communicating that, or is a bit of fuzziness harmless here?