Graphics : Improve search icons
- Reduce size by a couple of pixels.
- Make frame narrower than handle.
- Add glint on glass.
- Add `searchOff` variant to allow usage with _TogglePlugValueWidget.
- Add `searchFieldBackground` variant with additional padding to place it correctly as background image in search field.

Kept the old `search` icon name, as a copy of `searchOff` so that it doesn't disappear in a major version, but it is considered deprecated and will be removed in the future. I've updated this icon to match `searchOff` rather than kept it as it was - it seems more important that the size/layout matches `searchOn` than that it is exactly the same size as before.