SceneInspector : Improve filter UI
It was quite frustrating that you had to click on the search icon before being able to enter text in the search field - I kept finding myself clicking on the field directly and getting annoyed. It was also annoying that the placeholder text was doubly faded - once for being a placeholder and then again for being disabled.

Using `_TogglePlugValueWidget` allows us to enter text directly, with the toggle updating to reflect that fact.

This does mean replacing the OptionalValuePlug with a straight StringPlug, which strictly speaking might be a compatibility break. But we're giving ourselves a break here because :

- It's just the UI, and isn't serialised anywhere.
- The plug is brand spanking new, and we really doubt that anyone has scripted anything to set it up programmitically.