GafferSceneUI : Add VisibilityColumn
This is a specialised InspectorColumn that presents the true visibility state of scene locations to the user, rather than the raw value of the `scene:visible` attribute. This makes it more apparent that hiding a location also hides all of its descendants no matter the value of their `scene:visible` attribute.