PathListingWidget : Update visible items before the rest
This gives us much faster update times when there are many items out of view (whether collapsed or scrolled).

We treat recursive expansion as high priority. This makes sense - the user has explicitly asked that we open everything, so that is more important than any other updates. This appears to fix a regression when re-expanding a large hierarchy that has been expanded before. I'm not exactly sure _why_, but it does seem logical to give this priority.