PathListingWidget : Improve recursive expansion performance
This is a substantial performance improvement when expanding ~1,000,000
locations which have previously been expanded and collapsed. A delay of 10s in which Gaffer is pretty much unresponsive is reduced to an almost immediate update.

Profiling the delay using `perf` showed that Qt was spending all its time in  `QTreeViewPrivate::insertViewItems`, which I presume to be updating its internal layout of items to display. I don't _know_, but my guess is that we were causing quadratic scaling by repeatedly triggering insertions at the top of `QTreeViewPrivate::viewItems` (a QList, which despite the name is like a `std::vector`), and shuffling down all the following items.

My understanding is that `scheduleDelayedItemsLayout()` causes the layout to be done once for each time round the event loop rather than once per call to `setExpanded()`. And since we are delivering a storm of calls to `setExpanded()` this substantially reduces overhead. That `scheduleDelayedItemsLayout()` is
intended for this sort of use case is reinforced by the following from `QTreeView::expandRecursively` :

```
// do layouting only once after expanding is done
d->doDelayedItemsLayout();
...
```