RenderMan Globals : Work around failure to switch camera
It appears that hdPrman works around this by only ever making a single camera, and just changing which source camera it uses to update it from. That could be an option for us too - with more work - but for now we use an easier workaround based on making a new render view. Hopefully one day `Riley::ModifyRenderCamera()` can be fixed to do what it says on the tin, and we can remove the workaround.