RenderMan Renderer : Allow specification of Riley Variant (XPU/RIS)
This is not exposed via `Renderer::registerType()` yet, so there is no way for a client to request the XPU variant. We'll add the registration only once we have all the required bits and bobs in place (some of which require us to be able to query the variant - hence this commit comes first).