Menu : Automatically create shortcuts for keys on the number pad
Qt differentiates keys on the keypad by adding the `KeypadModifier` to their keypresses. This means that shortcuts that don't specifically include `KeypadModifier`, such as "Alt+1", would not be triggered by `Alt`+`keypad 1`, while "Alt+Num+1" would be triggered by `Alt`+`keypad 1` but not `Alt`+`1`. We want keypad keys to be treated the same as regular keys, and not require creators of menu items to remember to create two separate shortcuts, so this PR automatically creates the keypad specific shortcut when necessary.