SetMembershipInspector : Don't hardcode context variable name
`setMembership:set` is just the name that happens to be used by EditScopeAlgo's SetMembershipEdits processor. It's highly unlikely to be used anywhere else, and we shouldn't even be aware of its existence. By getting the variable name from the `setVariable` plug we maintain compatibility with the SetMembershipEdits processors but gain compatibility with everything else.