Metadata : Remove duplicate keys in `registeredValues()`
Each of the maps we retrieve keys from preserved uniqueness internally, but there are no such guarantees when we combine keys from multiple maps. So when we detect that we have such a combination, we remove duplicates before returning the result. This fixes a bug whereby the GraphEditor would show multiple copies of an annotation if a default value for `annotation:<name>:text` was registered statically against a node type.

As mentioned in the comment, although this approach to deduplication is `O( N^2 )`, it works well in practice, beating a separate `unordered_set` for 100 keys, which I think is a plausible upper bound in practice. For 1000 keys, this approach is around 2x slower than `unordered_set`, which still isn't disastrous if we do encounter more keys than expected.

Even if number of keys becomes a concern then the algorithm might not be the first thing to reconsider anyway. Most clients of `registeredValues()` do a post-filtering of the keys using a match pattern - to find everything starting with `preset:` or `annotation:` for example. If the match pattern was instead passed to `registeredValues()` and the filtering done internally, the number of keys to deduplicate could be decreased dramatically.