Inspector : Remove `SourceType.Fallback`
This was obscuring the source of edits which deleted a property, meaning we failed to indicate the edit with the blue cell background in our various scene editors.

I've also added a `useFallbacks` argument to `Result::value()`, although strictly speaking it wasn't necessary for this change. I felt that it helped to clarify the semantics of `sourceType()` and `source()` to be able to talk about a value without fallbacks. We have also talked about adding an option to hide fallbacks in the AttributeEditor, and it would be needed for that.