MetadataBinding : Accepts dicts of plug metadata
We didn't use dicts before because they didn't guarantee iteration order, and metadata order is significant (with presets ordering being one important use case). But in Python 3.6, the dict implementation switched to one which maintained insertion order, and in Python 3.7 this was enshrined in the language spec rather than being a mere implementation detail :

https://docs.python.org/3/whatsnew/3.7.html
https://docs.python.org/3/reference/datamodel.html#dictionaries

Dictionaries are the most natural way of defining key/value metadata pairs, so it makes sense that we should now prefer them to lists. This also gives us much better error messages when we mess up the formatting, and syntax highlighting to help us not mess it up in the first place. Whereas with lists, one missing comma would lead to an odd number of entries and a less than helpful runtime error.