Arnold Renderer : Set `options.frame` Arnold parameter
This is needed to get the right value for the `arnold/frame` EXR metadata written by `driver_exr`, and also for the frame number in `imager_overlay`. The Arnold devs assure me that it is only used for this sort of informational purpose, and that it will have no effect on images themselves (it's not related to motion blur in the same way that `reference_time` is).

The motivating factor here is that `noice` doesn't work properly with the EXR metadata, as reported on [gaffer-dev](https://groups.google.com/g/gaffer-dev/c/nAnRyHTLCUI/m/tyd5GBTuAQAJ).