Contrib : Add script to assist in converting metadata registrations
This has a completely brain dead approach to parsing, which is just barely sufficient to convert from lists to dictionaries in `registerMetadata()` calls. It's backed up by the ability to compare before/after metadata to find any discrepancies.

I've found it to be more reliable than doing it by hand myself (and much faster). But I don't trust it entirely - I recommend a good human eyeballing of all output.