RenderPassEditor : Prefer `columnLayout:label` metadata
We prefer this to the manually specified column names in `GafferSceneUI.RenderPassEditor.registerOption()` as we can reuse them when creating a column from a favourite. This also gets us closer to being able to populate the RenderPassEditor entirely from metadata.

Note that I haven't registered `columnLayout:label` metadata for the Arnold sampling and ray depth options and have left the manually specified column names. The manually specified column names remove the redundant "Depth" and "Samples" suffixes from the columns, which make sense given that those columns are registered within "Sampling" and "Ray Depth" sections, but when adding these columns to a more generic "Favourites" section it's better to keep the suffixes...