SceneInspector : Improve filtering
Previously we only matched the filter against locations where there was actually an inspector, meaning that it was impossible to match against primitive variable names (which are at intermediate levels in the hierarchy). We now use a hierarchical path filter as in the HierarchyView and AttributeEditor, allowing any part of the hierarchy to be matched.

I also experimented with a more complex approach whereby individual inspectors could be tagged as filterable or not. This had a few potential benefits :

- Searching for `P` wouldn't match the whole `Primitive Variables` section,
  because that would be tagged as non-filterable.
- Searching for `Interpolation` wouldn't match underneath every primitive variable, on the assumption that users wouldn't want to filter those.

This would have been combined with a "power" mode where including "/" in the filter string would perform full patch matching against everything instead. But having implemented it, I felt that it was confusing, due to it being really unclear why certain locations where matchable but not others, and the behaviour being different to our other editors. It puts us in a position where we have to read the users mind as to what they might be looking for.

So I went for this simpler and more predictable version instead.