ArnoldTextureBake : Remove `isolated` plug
We connect the `ArnoldTextureBake.dispatcher` plug to the
`__RenderDispatcher.dispatcher` plug. But the latter is a
`PythonCommand` which now gets a `dispatcher.isolated` plug added to it.
This causes the connection to error because
`ArnoldTextureBake.dispatcher` doesn't have `isolated`.
`__RenderDispatcher` is performing an Arnold render, so it makes sense
to remove this plug from it.