SelectionTool config : Indent USD kinds according to hierarchy
This is an attempt to make the hierarchy obvious without resorting to a bunch of cumbersome submenus. Eric did mention that he tried this in https://github.com/GafferHQ/gaffer/pull/5720 but I don't recall if we ever saw it during review. It seems kindof useful to me, so resurrecting it in an attempt to atone for the change of ordering in the previous commit.