RenderPassPlugValueWidget : Use metadata as single source of truth
My initial motivation for this was trying to help someone track down a problem that we suspected was related to creating gigantic menus - because they had a lot of passes and grouping was defaulted to off. This was in the SceneInspector's new diff widget, and it wasn't possible to change the menu configuration after the SceneInspector was already created because the metadata values were copied into member variables on construction. Using the metadata as a single source of truth fixes this, and seems like good practice in general.

There are a couple of things worth noting about this though :

- The state is now shared between the one in the menu bar and the one in the Settings window (because they are for the same plug).
- The menu state is now serialised with the `.gfr` file, and in future will be serialised into the SceneInspector's layout. If preferred, we could avoid this by making non-persistent metadata.

These both seem reasonable to me, but perhaps there was good reason for the original approach?