RenderPassPlugValueWidget : Refactor `self.__renderPasses` data structure
The previous representation had a few downsides :

- Every client had to do `__renderPasses.get( ..., [] )` in case a particular field hadn't been filled.
- It encouraged quadratic behaviour, where we did a loop over "all", and within the loop did linear search in "enabled" or "adaptorDisabled".
- It couldn't easily be extended to storing other properties of a render pass.

We solve these problems by storing passes in a dictionary mapping from pass name to the status for that pass. This surfaces the fact that we previously had no representation for passes which were created or enabled in an adaptor. I've dealt with that in the description for the pass, but not yet added an icon. That would need to be combined with changes to the main RenderPassEditor as well, so I'm calling out of scope for this fix-the-menu branch. But I do think adaptors creating passes is a valid use case, so we should potentially consider it for future work.