RenderPassEditor : Fix error handling and refresh
We were allowing `self.__renderPasses` to be `None` in four situations :

1. When an error occurred in `_valuesForUpdate()`.
2. When we had no plugs.
3. When we had multiple plugs and they disagreed on the current value.
4. When a background update had started and wasn't complete.

When in this state, `__menuDefinition()` would error because it used `__renderPasses` as a dictionary _before_ checking to see if it was `None`.
We now avoid the `None` state for `__renderPasses`, always at least having an empty dictionary.

The intention behind the `None` value appears to have been to trigger the creation of a "Refresh" menu item, but that is only relevant in situation 4. So we now use a separate `__updatePending` field to trigger that.