VectorDataWidget : Fix errors showing ColorChooser popups
When `_ColorDelegate` did `popup( parent = self )` it wasn't passing a widget as the parent - it was passing a QStyledItemDelegate. This caused later errors when the parent was retrieved and it was of an unexpected type.

It makes more sense to move the window parenting/popupping outside of `_createEditorInternal()` anyway, since the responsibility for setting the geometry for non-window editors is also outside of `_createEditorInternal()`. Now `_createEditorInternal()` and `editSignal()` are responsible only for what they are documented to be responsible for - creating the widget. Positioning and showing it are now handled automatically.

One way of demonstrating the previous bug is to create a RandomChoice node with Color3f type, and double click a swatch in the `Choices` widget to edit the colour.