RenderPassesUI : Show name editors in PopupWindow
Our default editor embedded neatly into the row of the VectorDataWidget, but that is unlikely to be the case for any custom widgets registered via `registerRenderPassNameWidget()`, which may well need more space. Rather than try to be clever and invent some heuristics to determine when the widget is small enough to embed, we simply show them all in popups. This has some precedent - spreadsheet row names are edited this way too.