PopupWindow : Fix problems when used as VectorDataWidget editor
If we close the window before the edit is committed, Qt outputs this message :

```
QAbstractItemView::commitData called with an editor that does not belong to this view
```

This could be seen when hitting Return in one of the numeric fields of a popup colour picker in a VectorDataWidget (a RandomChoice node is an easy way to test). Qt does it's own auto-close for editors anyway, so we can just skip ours.

Closing later triggers another problem though - an expired WeakMethod for the `paintEvent`. We avoid that by using a regular function instead of a method - we didn't need access to `self` anyway.

Not adding an entry to Changes.md because the colour picker didn't even show until the fix in 1c58086ab80b17471e41dce5d220fe679cbfe1a7 and I expect this commit to go out in the same release as that.