PlugValueWidgets : Fix usage of `_blockedUpdateFromValues()`
This context manager blocks the `plugDirtiedSignal` that is used to trigger updates in the widget. But when using an UndoScope, `plugDirtiedSignal` is emitted only when exiting the UndoScope. So the UndoScope needs to be nested inside `_blockedUpdateFromValues()` for us to get the effect we want.

Fixing this avoids unnecessary UI updates, with one very visible effect being that newly added rows in VectorDataPlugValueWidgets are no longer deselected as soon as they appear.