RenderMan Object : Fix default shadow group
Our RenderMan Light class is defaulting its `shadowSubset` parameter to "defaultShadowGroup", so for all objects to cast shadows by default, all objects need to be in that group. We were getting away with this in RIS, which seemed to treat ungrouped objects as matching anyway. But XPU is not so kind.