RenderMan Renderer : Add XPU device configuration options
I don't think there's a truly standard name for these, because they don't appear in any of the USD schemas published with RenderMan. In `hdPrman` they are specified as parameters in an HdRenderSettingsMap, and the USD documentation says they want to phase this out and put the render settings in scene description instead, by which I assume they mean moving them to schemas and storing the values on prims in USD. It seems unlikely that they'll use the native RenderMan `xpu:cpuconfig` name for that, so I've taken the `ri:xpuCpuConfig` name from `hdPrman`, in the hope that they'll keep the name the same when moving to a schema.

The GPU configuration is a little awkward - values seem to be just integers starting at zero. So the device type that `0` gives you on one machine may be different to the device that `0` gives you on another. I've avoided the temptation of doing something like IECoreCycles where we accept wildcards specifying GPU names, because I assume that interchange of configs (i.e. writing them to USD one day) is more important, and in practice not many folks are going to have more than one GPU.

I have deviated from RenderMan's native behaviour in another respect though : always specifying both parameters to `CreateRiley()`, seeding them with our own defaults. Otherwise it would be impossible to show a consistent fallback value in the Gaffer UI. Natively, RenderMan uses all devices if no parameters are specified, but stops using GPU devices as soon as a CPU config is provided (even though the former behaviour would suggest that GPUs default on). To make this more consistent, we default CPUs on and GPUs off.