_InspectorColumn : Remove reference to nonexistent enum value
This was removed in 91d305abfde482dc4231f9558c240565f602d461 (Gaffer 1.6.3.0). For the fix I've followed the presentation logic from `InspectorColumn::cellDataFromInspection()`, preferring a fallback description over a source plug. We might want to revisit at some point in the future though, as it's possible to have both (when the source was responsible for deleting the property).