SceneInspector : Fix handling of USD materials with nested UsdShadeNodeGraphs
Our own ShaderNetwork class doesn't have the concept of a nested subgraph, so when we load nested UsdShadeNodeGraphs from USD the container itself is omitted and the children are loaded with handles of the form `<subgraphPrimName>/<shaderPrimName>`. We therefore cannot use shader handles directly as a component within a `Gaffer::Path`, because `/` is invalid within a name. So we split the name on `/` to make as many path components as we need.