CustomAttributes, CustomOptions : Disable sectioning in plug layout
Otherwise we get inconsistent layouts when creating options or attributes which are registered with the metadata system (because they are part of StandardAttribute or CyclesOptions for example). In this case, the plug would appear in the main section when first created, but upon reopening the NodeEditor would move to a custom section as defined by the metadata.

Since the sections aren't useful in the context of the CustomOptions/CustomAttributes ndoes, we simply remove them rather than try to make them work consistently.