SConstruct : Add `-Wl,--as-needed` to GCC link flags
This causes libraries to be omitted if they do not provide any symbols needed by the target. This is actually a pain for us, because we have several libraries which don't provide symbols publicly but which do register visualisers or other UI components during initialisation. But that pain is unavoidable on Windows, and it's better to have the same problem on all platforms so we discover it sooner during development, and can deal with it more uniformly.

What prompted this switch was that we discovered that apparently the linker defaults to this behaviour on Ubuntu, so what we thought was a problem isolated to Windows was actually hitting some folks building on Linux.