Arnold Renderer : Make ArnoldOutputs immutable
The `ArnoldOutput::update()` function was added in 7223221d770a2a939038a7b2d570c629228517ea so that we could reuse driver nodes. This was needed because if we made new interactive drivers we would get new images in Gaffer's Catalogue instead of just updating the current one. But now driver node creation is deferred to allow us to share drivers between outputs, so we have no need to update outputs in place.

Storing immutable ArnoldOutputs is simpler and easier to reason about, so that's what we do now.