Arnold Renderer : Fix nondeterministic EXR part order
It was previously defined by the ordering of InternedStrings, which is not consistent from process to process. We now sort outputs so that RGBA comes first, and everything else comes alphabetically sorted after it.

Although Gaffer (and Nuke from the sounds of it) doesn't really care about part order, some apps (RV seems to be one) show the first part by default, causing confusion if that part keeps changing.