PlugValueWidget : Fix redundant updates
The ContextTracker emits `changedSignal()` every time it completes an update, but that doesn't mean that the tracked contexts have actually changed. So we now compare with the previous context to avoid doing redundant updates. This avoids some nasty flicker that was particularly noticeable in the `channels` and `view` widgets in the ImageView's toolbar, especially when focussed on the output of a running InteractiveRender.