ImageReader : Improve channel name heuristics for ID channel
When writing a multi-part EXR file containing both RGBA and `id` channels, Arnold will put the `id` channel in an image part called `id`. This gives us exactly what we want according to the EXR specification, but ran afoul of our heuristics in `ChannelInterpretation::Default` mode, which generally assume that channels should have a layer name.

Since `id` is a [standard EXR attribute](https://openexr.com/en/latest/DeepIDsSpecification.html#standard-id-channel-names), I believe that it shouldn't have a layer name, just like the other standard channels "R", "G", "B", "A" and "Z". And ImageSelectionTool was written with that belief too, since it looks for an `id` channel rather than `id.id`. So this commit adjust the heuristics to match those beliefs.

This also adjusts the heuristics for `instanceID` in the same way. Strictly speaking this is not a standard EXR channel name, because the spec has a lower-case `instanceid` instead. But `instanceID` is serving the same purpose, and we should adjust our casing to match at some point in the future.

Fixes #6678.