Multi-index containers : Prefer `boost::multi_index<key>`
This is a less verbose means of expressing keys that became available with C++ 17. It boils down to the original `member` or `mem_fn` through some template trickery behind the scenes. Since the verbosity of type definitions is perhaps the main downside of `multi_index_container`, this seems like a small but worthwhile improvement.