GridContainer : Add `nextRow()` context manager
Generally speaking, the great thing about using the `with` syntax for parenting widgets is that the structure of the UI layout is reflected in the structure of the code. But not so for GridContainer, where there's just a big old soup of child widgets and the structure is achieved by specifying indices manually. The `nextRow()` context manager is intended to fix this, by reflecting the creation of rows in the structure of the code, and automatically generating indices for you.