GafferSceneUI : Use `render:defaultRenderer` metadata as source of truth
This isn't quite as neat as we thought it might be, because everyone has their own idea of what the label should be for a value of "". But it does mean we only handle the `ui:enabled` renderer metadata in one place now. And it will reduce the number of spots we will need to consider an upcoming `label` metadata.

This also updates RenderPassShaderUI to use a more general method of avoiding duplicate entries for renderer variants (like 3Delight Cloud). This will work naturally with an upcoming RenderManXPU variant.