Delight Renderer : Add todo
We don't like spaces in names, because it makes it a pain to name plugs or nodes to match. This is an issue for the SceneView which wants to name a plug to match the renderer name. Not an issue for 3Delight Cloud as its not suitable for usage in the Viewer, but something we want to avoid in general.