IECoreArnold : Remove deprecated/dupe call to expandSplineParameters
IECoreScene::ShaderNetworkAlgo::expandSplineParameters was the deprecated way to handle spline parameters.

IECoreArnold::ShaderNetworkAlgo is now doing the up to date approach - there is a call to IECoreScene::ShaderNetworkAlgo::convertToOSLConventions in preprocessedNetwork, which will have already converted splines. As far as I can tell, this means that expandSplineParameters was currently not doing anything, and the fact it was still being called was simply an oversight.