SConstruct : Compress debug info
Debug builds are quite large and we're fighting the limited available space on GitHub's CI runners. Compressing the debug info is a fair win in reducing the overall size of a debug build. We use zlib compression as it is compatible with our current toolchain, once we go to VFX Platform 2026 with GCC 14, there's the potential for using zstd compression for further wins.