OptionsUI : Fix missing presets
Presets can be registered either via `presetNames` or `preset:{name}` metadata, but we were only considering the former. OptionTweaksUI got this right, by combining the two, so we base our implementation on that. Moved the implementation to OptionsUI as that seems to me the primary node, and shared it back to OptionTweaksUI to avoid duplication.

The only native preset I know of that was affected by this was "Standard" for `renderPass:type`. But now that this is fixed, I suspect we might prefer the `preset:{name}` syntax for more metadata registrations in future - it's a bit simpler than managing two separate lists.