RenderMan Globals : Don't use Loader during static initialisation
This fixes crashes when running the 26.3 tests on CI, following the introduction of the XPU tests. The trigger for the crash was a reordering of the `import` statements, with IECoreRenderMan now imported before IECoreUSD.
The static initialisation in IECoreRenderMan's Globals.cpp was calling `Loader::strings()` which meant we were opening `libprman.so` earlier than before. And unfortunately, PRMan's `libstats.so` exports symbols related to `std::regex` which are not compatible with our libraries. With IECoreUSD being imported after `libstats.so` it picked up these bad symbols, leading us to the following crash while importing IECoreUSD:

```
/opt/pixar/RenderManProServer-26.3/lib/../lib/libstats.so(_ZNSt8__detail8_ScannerIcE14_M_scan_normalEv+0x32)[0x7ff62fc3b442]
/__w/gaffer/gaffer/build/lib/libusd_tf.so(_ZNSt8__detail9_CompilerINSt7__cxx1112regex_traitsIcEEE11_M_try_charEv+0x16f)[0x7ff627cc966f]
/__w/gaffer/gaffer/build/lib/libusd_tf.so(_ZNSt8__detail9_CompilerINSt7__cxx1112regex_traitsIcEEE7_M_atomEv+0x22)[0x7ff627cd1b52]
/__w/gaffer/gaffer/build/lib/libusd_tf.so(_ZNSt8__detail9_CompilerINSt7__cxx1112regex_traitsIcEEE14_M_alternativeEv+0xf0)[0x7ff627cd2180]
/__w/gaffer/gaffer/build/lib/libusd_tf.so(_ZNSt8__detail9_CompilerINSt7__cxx1112regex_traitsIcEEE14_M_disjunctionEv+0x19)[0x7ff627cd2459]
/__w/gaffer/gaffer/build/lib/libusd_tf.so(_ZNSt8__detail9_CompilerINSt7__cxx1112regex_traitsIcEEEC1EPKcS6_RKSt6localeNSt15regex_constants18syntax_option_typeE+0x34e)[0x7ff627cd2bce]
/__w/gaffer/gaffer/build/lib/libIECoreUSD.so(_ZNSt8__detail13__compile_nfaINSt7__cxx1112regex_traitsIcEEPKcEENSt9enable_ifIXsrNS_20__is_contiguous_iterIT0_EE5valueESt10shared_ptrIKNS_4_NFAIT_EEEE4typeES8_S8_RKNSC_11locale_typeENSt15regex_constants18syntax_option_typeE+0x1f)[0x7ff60c39361f]
/__w/gaffer/gaffer/build/lib/libIECoreUSD.so(_ZNSt7__cxx1111basic_regexIcNS_12regex_traitsIcEEEC1EPKcNSt15regex_constants18syntax_option_typeE+0x56)[0x7ff60c3937b6]
/__w/gaffer/gaffer/build/lib/libIECoreUSD.so(+0xc141d)[0x7ff60c2c141d]
/lib64/ld-linux-x86-64.so.2(+0x8f1a)[0x7ff64cf06f1a]
/lib64/ld-linux-x86-64.so.2(+0x901a)[0x7ff64cf0701a]
/lib64/libc.so.6(_dl_catch_exception+0xdc)[0x7ff64b5f9c6c]
/lib64/ld-linux-x86-64.so.2(+0x1076e)[0x7ff64cf0e76e]
/lib64/libc.so.6(_dl_catch_exception+0x84)[0x7ff64b5f9c14]
/lib64/ld-linux-x86-64.so.2(+0x10951)[0x7ff64cf0e951]
/lib64/libdl.so.2(+0xf8a)[0x7ff64bddcf8a]
/lib64/libc.so.6(_dl_catch_exception+0x84)[0x7ff64b5f9c14]
/lib64/libc.so.6(_dl_catch_error+0x33)[0x7ff64b5f9cd3]
/lib64/libdl.so.2(+0x152e)[0x7ff64bddd52e]
/lib64/libdl.so.2(dlopen+0x4a)[0x7ff64bddd02a]
/__w/gaffer/gaffer/build/lib/libpython3.11.so.1.0(+0x2b7e66)[0x7ff64c4b7e66]
```

I'm unable to prove this, but this bug report suggests that the symbols being exported from PRMan are not using the C++11 ABI, while our libraries are :

https://gcc.gnu.org/bugzilla/show_bug.cgi?id=118408

Delaying IECoreRenderMan's usage of the Loader means that we pick up the required symbols from compatible libraries rather than PRMan's.

We will be forever vulnerable to problems like this for as long as we need to `dlopen()` `libprman` with `RTLD_GLOBAL`, but all attempts at removing that (including using `RTLD_DEEPBIND`) lead to crashes in RenderMan itself, so that doesn't seem to be an option. It is worth noting that `hdPrman` uses `RTLD_GLOBAL` too, so this does seem to be the intended usage. So the best we can do right now is to delay loading `libprman.so` until the last possible moment, when we first create a Renderer instance.