Dispatcher compatibility : Remove unnecessary `dispatch()` restrictions
These were emulating restrictions from the original `Dispatcher::dispatch()` method that the config provides compatibility for. But they prevented more modern dispatch idioms such as using Switch or ContextVariables set up with TaskPlugs.

It's a little strange to be improving a compatibility config to add new features, but it's the best way of improving the `dispatch` app and DispatcherDialogue without introducing a breaking change. Added todos documenting how we should deal with this when we get the chance to make a break.