ReferenceUI : Fix order of operations for "Duplicate as Box"
Previously we were parenting and selecting the Box before loading its contents.
This meant that the parent's `childAddedSignal()` and the selection's `memberAddedSignal()`
were emitted when the node was in an intermediate state. We don't want to emit signals in
such a state, as we want client code to always see a final and consistent state. So now
we prepare the box in a temporary script and move it over once it is finalised.