ShaderTweaksUI : Improve creation of tweaks via SceneInspector drag
We now use `SceneInspector.draggedParameters()` to provide a consistent dict of parameter names and values when either a name or value is dragged from the Scene Inspector. This did require extending `draggedParameters()` to return disambiguated parameter names for non-output shaders, which was slightly more of a faff than I'd hoped.

With this we now support creating multiple tweaks when multiple parameters are dragged from the Scene Inspector. We're not doing any checking here to prevent the same tweaks from being created again if they already exist, but that would be straightforward to support...